fix n type

+++ src/lib.rs
@@ -31,8 +31,8 @@ pub fn get_bitonic_sequence(n: usize, start: i32, end: i32) -> Vec<i32> {
     my_array_solution.into_iter().collect()
 }
 
-pub fn get_bitonic_sequence_v2(n: i32, l: i32, r: i32) -> Vec<i32> {
-  if n > (r - l) * 2 + 1 {
+pub fn get_bitonic_sequence_v2(n: u32, l: i32, r: i32) -> Vec<i32> {
+  if n > ((r - l) * 2 + 1) as u32 {
     return [-1].to_vec();
   }
